requirements.txt
tkinter
pyperclip
cryptographyns
os
json
secrets
string
re